refactor: change Loading story metadata object
