+++ apps/web/src/components/TrendingItemSuggestion/TrendingItemSuggestion.test.tsx
@@ -1,6 +1,6 @@
 import { screen, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { metaData } from './TISmetaData';
+import { metaData } from './TrendingItemSuggestion.utils.tsx';
 import TrendingItemSuggestion from './TrendingItemSuggestion';
 
 describe('TrendingItemSuggestion', () => {
